requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
jinja2
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic
google-generativeai
